Feature #16: feat chip filter chip
